Soix/requirements.txt
pynput
discord
psutil
requests
pillow
discord-pretty-help
sounddevice
wavio
pyinstaller
pygame
configparser